Changed red / green / yellow variable names.
The new ones aren't that pretty, but I don't like to couple variable names to specific colours. The code should not state whether the bar is "yellow", "green", "red" or "pink", but it should reflect what the state is that will be shown. For the same reason, you should never call CSS classes "red", "bold" etc., since they are marking what kind of element should be displayed, not how it will look like,